requirements.txt
google-generativeai
pandas
openpyxl
python-dotenv

